+++ vite.config.ts
@@ -17,8 +17,8 @@ export default defineConfig({
     },
     rollupOptions: {
       external: [
+        // /^@lit.*/, // @lit/react패키지 빌드에 포함
         /^lit.*/,
-        /^@lit.*/,
         /^react.*/,
       ]
     },